@@ -25,7 +25,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 上级 
 		/// </summary>
-		public DeptViewModel Super {get;set;}
+		public string Super_Value {get;set;}
 		
 		/// <summary>
 		/// 部门代码 
@@ -38,6 +38,7 @@ namespace FastFrame.Application.Basis
 		/// </summary>
 		[StringLength(50)]
 		[Required()]
+		[IsPrimaryField()]
 		public string Name {get;set;}
 		
 		/// <summary>
@@ -56,7 +57,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 创建人 
 		/// </summary>
-		public UserViewModel Create_User {get;set;}
+		public string Create_User_Value {get;set;}
 		
 		/// <summary>
 		/// 创建时间 
@@ -75,7 +76,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 修改人 
 		/// </summary>
-		public UserViewModel Modify_User {get;set;}
+		public string Modify_User_Value {get;set;}
 		
 		/// <summary>
 		/// 修改时间 
