@@ -31,12 +31,13 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 上级值 
 		/// </summary>
-		public EnumItemViewModel Super {get;set;}
+		public string Super_Value {get;set;}
 		
 		/// <summary>
 		/// 树状码 
 		/// </summary>
 		[StringLength(20)]
+		[IsPrimaryField()]
 		public string TreeCode {get;set;}
 		
 		/// <summary>
@@ -66,7 +67,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 创建人 
 		/// </summary>
-		public UserViewModel Create_User {get;set;}
+		public string Create_User_Value {get;set;}
 		
 		/// <summary>
 		/// 创建时间 
@@ -85,7 +86,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 修改人 
 		/// </summary>
-		public UserViewModel Modify_User {get;set;}
+		public string Modify_User_Value {get;set;}
 		
 		/// <summary>
 		/// 修改时间 
