@@ -21,6 +21,7 @@ namespace FastFrame.Application.Basis
 		/// </summary>
 		[StringLength(50)]
 		[Required()]
+		[IsPrimaryField()]
 		public string EnCode {get;set;}
 		
 		/// <summary>
@@ -56,7 +57,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 创建人 
 		/// </summary>
-		public UserViewModel Create_User {get;set;}
+		public string Create_User_Value {get;set;}
 		
 		/// <summary>
 		/// 创建时间 
@@ -75,7 +76,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 修改人 
 		/// </summary>
-		public UserViewModel Modify_User {get;set;}
+		public string Modify_User_Value {get;set;}
 		
 		/// <summary>
 		/// 修改时间 
