@@ -1,4 +1,5 @@
 ﻿using FastFrame.Entity.Basis;
+using Microsoft.Extensions.DependencyInjection;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -9,7 +10,8 @@ namespace FastFrame.Application.Basis
         protected override async Task OnGeting(RoleDto dto)
         {
             await base.OnGeting(dto);
-            dto.Members = await EventBus.RequestAsync<UserViewModel[], RoleDto>(dto);
+            var user_dic = await Loader.GetService<RoleMemberService>().GetUserViewModelsByRoleIds(dto.Id);
+            dto.Members = user_dic.SelectMany(v => v.Value);
             dto.Permissions = (await EventBus.RequestAsync<RolePermission[], RoleDto>(dto)).Select(v => v.PermissionKey);
         }
     }
