@@ -1,12 +1,13 @@
-﻿using System.Collections.Generic;
+﻿using FastFrame.Entity;
+using System.Collections.Generic;
 
 namespace FastFrame.Application.Basis
 {
     public partial class UserDto
     {
-        public IEnumerable<DeptViewModel> Depts { get; set; }
+        public IEnumerable<IViewModel> Depts { get; set; }
 
-        public IEnumerable<RoleViewModel> Roles { get; set; }
+        public IEnumerable<IViewModel> Roles { get; set; }
     }
 
 }
