@@ -21,6 +21,7 @@ namespace FastFrame.Application.Basis
 		/// </summary>
 		[StringLength(50)]
 		[Required()]
+		[IsPrimaryField()]
 		public string Account {get;set;}
 		
 		/// <summary>
@@ -61,7 +62,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 头像 
 		/// </summary>
-		public ResourceViewModel HandIcon {get;set;}
+		public string HandIcon_Value {get;set;}
 		
 		/// <summary>
 		/// 是否管理员 
@@ -83,7 +84,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 创建人 
 		/// </summary>
-		public UserViewModel Create_User {get;set;}
+		public string Create_User_Value {get;set;}
 		
 		/// <summary>
 		/// 创建时间 
@@ -102,7 +103,7 @@ namespace FastFrame.Application.Basis
 		/// <summary>
 		/// 修改人 
 		/// </summary>
-		public UserViewModel Modify_User {get;set;}
+		public string Modify_User_Value {get;set;}
 		
 		/// <summary>
 		/// 修改时间 
