@@ -25,9 +25,9 @@ namespace FastFrame.Application.Flow
 			this.dFModuleRepository=dFModuleRepository;
 		}
 		
-		protected override IQueryable<DFModuleDto> QueryMain() 
+		protected override IQueryable<DFModuleDto> DefaultQueryable() 
 		{
-			var userQueryable = userRepository.Queryable.MapTo<User,UserViewModel>();
+			var userQueryable = userRepository.Queryable.Select(User.BuildExpression());
 			var repository = dFModuleRepository.Queryable;
 			var query = from _dFModule in repository 
 						join _create_User_Id in userQueryable on _dFModule.Create_User_Id equals _create_User_Id.Id into t__create_User_Id
@@ -47,15 +47,10 @@ namespace FastFrame.Application.Flow
 							CreateTime = _dFModule.CreateTime,
 							Modify_User_Id = _dFModule.Modify_User_Id,
 							ModifyTime = _dFModule.ModifyTime,
-							Create_User = _create_User_Id,
-							Modify_User = _modify_User_Id,
+							Create_User_Value = _create_User_Id.Value,
+							Modify_User_Value = _modify_User_Id.Value,
 						};
 			return query;
-		}
-		public Task<IPageList<DFModuleViewModel>> ViewModelListAsync(IPagination<DFModuleViewModel> page) 
-		{
-			var query = dFModuleRepository.MapTo<DFModule, DFModuleViewModel>();
-			return query.PageListAsync(page);
 		}
 		
 	}
