@@ -27,15 +27,18 @@ namespace FastFrame.Application.OA
 			this.oaLeaveRepository=oaLeaveRepository;
 		}
 		
-		protected override IQueryable<OaLeaveDto> QueryMain() 
+		protected override IQueryable<OaLeaveDto> DefaultQueryable() 
 		{
-			var userQueryable = userRepository.Queryable.MapTo<User,UserViewModel>();
-			var deptQueryable = deptRepository.Queryable.MapTo<Dept,DeptViewModel>();
+			var userQueryable = userRepository.Queryable.Select(User.BuildExpression());
+			var deptQueryable = deptRepository.Queryable.Select(Dept.BuildExpression());
 			var repository = oaLeaveRepository.Queryable;
 			var query = from _oaLeave in repository 
-						join _create_User_Id in userQueryable on _oaLeave.Create_User_Id equals _create_User_Id.Id 
-						join _dept_Id in deptQueryable on _oaLeave.Dept_Id equals _dept_Id.Id 
-						join _agent_Id in userQueryable on _oaLeave.Agent_Id equals _agent_Id.Id 
+						join _create_User_Id in userQueryable on _oaLeave.Create_User_Id equals _create_User_Id.Id into t__create_User_Id
+						from _create_User_Id in t__create_User_Id.DefaultIfEmpty()
+						join _dept_Id in deptQueryable on _oaLeave.Dept_Id equals _dept_Id.Id into t__dept_Id
+						from _dept_Id in t__dept_Id.DefaultIfEmpty()
+						join _agent_Id in userQueryable on _oaLeave.Agent_Id equals _agent_Id.Id into t__agent_Id
+						from _agent_Id in t__agent_Id.DefaultIfEmpty()
 						join _modify_User_Id in userQueryable on _oaLeave.Modify_User_Id equals _modify_User_Id.Id into t__modify_User_Id
 						from _modify_User_Id in t__modify_User_Id.DefaultIfEmpty()
 						select new OaLeaveDto
@@ -55,17 +58,12 @@ namespace FastFrame.Application.OA
 							Id = _oaLeave.Id,
 							Modify_User_Id = _oaLeave.Modify_User_Id,
 							ModifyTime = _oaLeave.ModifyTime,
-							Create_User = _create_User_Id,
-							Dept = _dept_Id,
-							Agent = _agent_Id,
-							Modify_User = _modify_User_Id,
+							Create_User_Value = _create_User_Id.Value,
+							Dept_Value = _dept_Id.Value,
+							Agent_Value = _agent_Id.Value,
+							Modify_User_Value = _modify_User_Id.Value,
 						};
 			return query;
-		}
-		public Task<IPageList<OaLeaveViewModel>> ViewModelListAsync(IPagination<OaLeaveViewModel> page) 
-		{
-			var query = oaLeaveRepository.MapTo<OaLeave, OaLeaveViewModel>();
-			return query.PageListAsync(page);
 		}
 		
 	}
