@@ -1,19 +1,19 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Linq.Expressions;
 
 namespace FastFrame.Entity.Basis
 {
     /// <summary>
     /// 部门
     /// </summary>
-    [Export]
-    [RelatedField(nameof(Name))]
-    public class Dept : BaseEntity, ITreeEntity
-    { 
+    [Export] 
+    public class Dept : BaseEntity, ITreeEntity, IViewModelable<Dept>
+    {
         /// <summary>
         /// 上级
         /// </summary>
         [RelatedTo(typeof(Dept))]
-        public string Super_Id { get; set; } 
+        public string Super_Id { get; set; }
 
         /// <summary>
         /// 部门代码
@@ -26,6 +26,7 @@ namespace FastFrame.Entity.Basis
         /// 部门名称
         /// </summary>
         [StringLength(50), Required]
+        [IsPrimaryField]
         public string Name { get; set; }
 
         /// <summary>
@@ -34,6 +35,13 @@ namespace FastFrame.Entity.Basis
         [StringLength(200)]
         public string Remarks { get; set; }
 
+
+        private static Expression<Func<Dept, IViewModel>> vm_expression = v => new DefaultViewModel { Id = v.Id, Value = v.Name + "(" + v.TreeCode + ")" };
+
+        public static Expression<Func<Dept, IViewModel>> BuildExpression()
+        {
+            return vm_expression;
+        }
 
         public void SetNumber(string val)
         {
