@@ -1,14 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Linq.Expressions;
 
 namespace FastFrame.Entity.Basis
 {
     /// <summary>
     /// 图片库
-    /// </summary> 
-    [RelatedField(nameof(Name))]
+    /// </summary>  
     [Unique(nameof(Super_Id), nameof(Name))]
-    [Export(ExportMark.Service, ExportMark.DTO, ExportMark.ViewModel)]
-    public class Meidia : BaseEntity,ITreeEntity
+    [Export(ExportMark.Service, ExportMark.DTO)]
+    public class Meidia : BaseEntity, ITreeEntity, IViewModelable<Meidia>
     {
         /// <summary>
         /// 上级
@@ -21,12 +21,13 @@ namespace FastFrame.Entity.Basis
         /// </summary>
         [StringLength(50)]
         [Required]
+        [IsPrimaryField]
         public string Name { get; set; }
 
         /// <summary>
         /// 资源
         /// </summary> 
-        public string Resource_Id { get; set; } 
+        public string Resource_Id { get; set; }
 
         /// <summary>
         /// 是否文件夹
@@ -38,7 +39,7 @@ namespace FastFrame.Entity.Basis
         /// </summary>
         [ReadOnly]
         [StringLength(200)]
-        public string TreeCode { get; set; } 
+        public string TreeCode { get; set; }
 
         public void SetNumber(string val)
         {
@@ -46,5 +47,13 @@ namespace FastFrame.Entity.Basis
         }
 
         public string GetNumber() => TreeCode;
+
+
+        private static Expression<Func<Meidia, IViewModel>> vm_expression = v => new DefaultViewModel { Id = v.Id, Value = v.Name + "(" + v.TreeCode + ")" };
+
+        public static Expression<Func<Meidia, IViewModel>> BuildExpression()
+        {
+            return vm_expression;
+        }
     }
 }
