@@ -1,18 +1,19 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Linq.Expressions;
 
 namespace FastFrame.Entity.Basis
 {
     /// <summary>
     /// 角色
-    /// </summary>
-    [RelatedField(nameof(Name), nameof(EnCode))]
+    /// </summary> 
     [Export]
-    public class Role : BaseEntity
+    public class Role : BaseEntity, IViewModelable<Role>
     {
         /// <summary>
         /// 编码
         /// </summary>
         [StringLength(50), Required, Unique]
+        [IsPrimaryField]
         public string EnCode { get; set; }
 
         /// <summary>
@@ -29,12 +30,17 @@ namespace FastFrame.Entity.Basis
         /// <summary>
         /// 管理员角色
         /// </summary>
-        public bool IsAdmin { get; set; } 
+        public bool IsAdmin { get; set; }
 
         /// <summary>
         /// 备注
         /// </summary>
         [StringLength(500)]
         public string Remarks { get; set; }
+
+
+        private static Expression<Func<Role, IViewModel>> vm_expression = v => new DefaultViewModel { Id = v.Id, Value = v.Name + "(" + v.EnCode + ")" };
+
+        public static Expression<Func<Role, IViewModel>> BuildExpression() => vm_expression;
     }
 }
