@@ -6,12 +6,12 @@
     public class RoleMember : TableMap
     {
         /// <summary>
-        /// 角色
+        /// 角色 <see cref="Role"/>
         /// </summary> 
         public override string FKey_Id { get; set; }
 
         /// <summary>
-        /// 用户
+        /// 用户 <see cref="User"/>
         /// </summary> 
         public override string Value_Id { get; set; } 
     }
