@@ -1,5 +1,7 @@
 ﻿using FastFrame.Application;
 using FastFrame.Application.Basis;
+using FastFrame.Entity;
+using FastFrame.Entity.Basis;
 using FastFrame.Infrastructure;
 using FastFrame.Infrastructure.Permission;
 using Microsoft.AspNetCore.Mvc;
@@ -12,8 +14,7 @@ namespace FastFrame.WebHost.Controllers.Basis
     {
         [Permission(new string[] { "Add", "Update" })]
         [HttpGet]
-        public Task<IPageList<UserViewModel>> UserList(string qs)
-            => Request.HttpContext.RequestServices
-                    .GetService<UserService>().ViewModelListAsync(Pagination<UserViewModel>.FromJson(qs)); 
+        public Task<IEnumerable<IViewModel>> UserList(string kw, int page_index = 1, int page_size = 10)
+            => service.ViewModelListAsync<User>(kw, page_index, page_size);
     }
 }
